+++ src/main/java/abu/lets_play/Model/dto/ProductPatchRequest.java
@@ -3,9 +3,9 @@ package abu.lets_play.Model.dto;
 public class ProductPatchRequest {
     private String name;
     private String description;
-    private Integer price;
+    private Double price;
 
     public String getName() { return name; }
     public String getDescription() { return description; }
-    public Integer getPrice() { return price; }
+    public Double getPrice() { return price; }
 }
